Premium calculation uses int

+++ src/vote.cpp
@@ -446,8 +446,8 @@ uint64 GetPremium(uint64 nValue, uint64 nDuration, unsigned char cUnit, const st
 
                 const CParkRate& prevParkRate = vSortedParkRate[i-1];
 
-                double ratio = (double)(parkRate.nRate - prevParkRate.nRate) / (parkRate.GetDuration() - prevParkRate.GetDuration());
-                double rate = prevParkRate.nRate + (nDuration - prevParkRate.GetDuration()) * ratio;
+                int ratio = (int)(parkRate.nRate - prevParkRate.nRate) / (parkRate.GetDuration() - prevParkRate.GetDuration());
+                int rate = prevParkRate.nRate + (nDuration - prevParkRate.GetDuration()) * ratio;
                 return nValue * rate / COIN;
             }
         }
